DIOS-3710 Fix duplicate timestamps in transcoding (#250)
* Using a queue we prevent GrabFrame() from getting the same frame more than once fixing the duplicate timestamp issue
* Changes VideoPipe to use a queue of size 2 instead of double buffering (can reduce latency in some circumstances)
* Adds a log message so we can identify when frames are being dropped

+++ include/video.h
@@ -158,7 +158,9 @@ public:
 	virtual ~VideoOutput() = default;
 
 	virtual void ClearFrame() = 0;
-	virtual int NextFrame(const VideoBuffer::const_shared& videoBuffer)=0;
+
+	// Returns the current occupancy of the frame buffer queue
+	virtual size_t NextFrame(const VideoBuffer::const_shared& videoBuffer)=0;
 };
